+++ src/main/java/de/samply/file/bundle/PathsBundleManager.java
@@ -1,11 +1,15 @@
 package de.samply.file.bundle;
 
+import java.nio.file.Path;
+
 public interface PathsBundleManager {
 
   boolean isNextPathsBundleInInputFolder() throws PathsBundleManagerException;
 
   PathsBundle fetchNextPathsBundleFromInputFolder() throws PathsBundleManagerException;
 
-  void movePathsBundleToOutputFolder(PathsBundle pathsBundle) throws PathsBundleManagerException;
+  void movePathsBundleToOutputFolder(PathsBundle pathsBundle, Path outputFolderPath) throws PathsBundleManagerException;
+
+  PathsBundle copyPathsBundleToOutputFolder(PathsBundle pathsBundle, Path outputFolderPath) throws PathsBundleManagerException;
 
 }